backend/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pandas==2.2.3
python-multipart==0.0.17
pytest==8.3.4
openpyxl==3.1.5
anthropic>=0.40.0
claude-agent-sdk
python-dotenv>=1.0.1
sqlalchemy>=2.0
psycopg2-binary
alembic
passlib[bcrypt]
bcrypt
python-jose[cryptography]
pyyaml>=6.0
